+++ src/main/java/com/example/bunsanedthinking_springback/model/domain/contract/ContractDModel.java
@@ -1,12 +1,5 @@
 package com.example.bunsanedthinking_springback.model.domain.contract;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Optional;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
 import com.example.bunsanedthinking_springback.entity.compensationDetail.CompensationDetail;
 import com.example.bunsanedthinking_springback.entity.contract.Contract;
 import com.example.bunsanedthinking_springback.entity.depositDetail.DepositDetail;
@@ -20,6 +13,12 @@ import com.example.bunsanedthinking_springback.model.domain.paymentDetail.Paymen
 import com.example.bunsanedthinking_springback.model.domain.product.ProductDModel;
 import com.example.bunsanedthinking_springback.repository.ContractMapper;
 import com.example.bunsanedthinking_springback.vo.ContractVO;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Optional;
 
 @Service
 public class ContractDModel {
@@ -72,15 +71,68 @@ public class ContractDModel {
 		return contractMapper.getMaxId_SalesModel();
 	}
 
-	public void add(ContractVO contractVO) {
-		contractMapper.insert_SalesModel(contractVO);
+	public void add(Contract contract) {
+		if (contract == null) return;
+		if (contractMapper.getById_Customer(contract.getId()).isPresent()) return;
+		contractMapper.insert_SalesModel(contract.getVO());
+
+		List<InsuranceMoney> insuranceMonies = contract.getInsuranceMoneyList();
+		if (insuranceMonies != null) insuranceMonies.forEach(e -> insuranceMoneyDModel.add(e));
+
+		List<CompensationDetail> compensationDetails = contract.getCompensationDetailList();
+		if (insuranceMonies != null) compensationDetails.forEach(e -> compensationDetailDModel.add(e));
+
+		List<DepositDetail> depositDetails = contract.getDepositDetailList();
+		if (depositDetails != null) depositDetails.forEach(e -> depositDetailDModel.add(e));
+
+		List<PaymentDetail> paymentDetails = contract.getPaymentDetailList();
+		if (paymentDetails != null) paymentDetails.forEach(e -> paymentDetailDModel.add(e));
+
+		Product product = contract.getProduct();
+		if (product != null) productDModel.add(product);
 	}
 
-	public void update(ContractVO contractVO) {
-		contractMapper.update(contractVO);
+	public void update(Contract contract) {
+		if (contract == null) return;
+		if (contractMapper.getById_Customer(contract.getId()).isEmpty()) return;
+
+		List<InsuranceMoney> insuranceMonies = contract.getInsuranceMoneyList();
+		if (insuranceMonies != null) insuranceMonies.forEach(e -> insuranceMoneyDModel.update(e));
+
+		List<CompensationDetail> compensationDetails = contract.getCompensationDetailList();
+		if (insuranceMonies != null) compensationDetails.forEach(e -> compensationDetailDModel.update(e));
+
+		List<DepositDetail> depositDetails = contract.getDepositDetailList();
+		if (depositDetails != null) depositDetails.forEach(e -> depositDetailDModel.update(e));
+
+		List<PaymentDetail> paymentDetails = contract.getPaymentDetailList();
+		if (paymentDetails != null) paymentDetails.forEach(e -> paymentDetailDModel.update(e));
+
+		Product product = contract.getProduct();
+		if (product != null) productDModel.update(product);
+
+		contractMapper.update(contract.getVO());
 	}
 
 	public void delete(int id) {
+		if (contractMapper.getById_Customer(id).isEmpty()) return;
+		Contract contract = getById(id);
+
+		List<InsuranceMoney> insuranceMonies = contract.getInsuranceMoneyList();
+		if (insuranceMonies != null) insuranceMonies.forEach(e -> insuranceMoneyDModel.delete(e.getId()));
+
+		List<CompensationDetail> compensationDetails = contract.getCompensationDetailList();
+		if (insuranceMonies != null) compensationDetails.forEach(e -> compensationDetailDModel.delete(e.getId()));
+
+		List<DepositDetail> depositDetails = contract.getDepositDetailList();
+		if (depositDetails != null) depositDetails.forEach(e -> depositDetailDModel.delete(e.getId()));
+
+		List<PaymentDetail> paymentDetails = contract.getPaymentDetailList();
+		if (paymentDetails != null) paymentDetails.forEach(e -> paymentDetailDModel.delete(e.getId()));
+
+		Product product = contract.getProduct();
+		if (product != null) productDModel.delete(product.getId());
+
 		contractMapper.deleteById(id);
 	}
 }
