+++ src/main/java/com/example/bunsanedthinking_springback/model/domain/diseaseHistory/DiseaseHistoryDModel.java
@@ -1,14 +1,13 @@
 package com.example.bunsanedthinking_springback.model.domain.diseaseHistory;
 
-import java.util.ArrayList;
-import java.util.List;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
 import com.example.bunsanedthinking_springback.entity.diseaseHistory.DiseaseHistory;
 import com.example.bunsanedthinking_springback.repository.DiseaseHistoryMapper;
 import com.example.bunsanedthinking_springback.vo.DiseaseHistoryVO;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import java.util.ArrayList;
+import java.util.List;
 
 @Service
 public class DiseaseHistoryDModel {
@@ -33,15 +32,16 @@ public class DiseaseHistoryDModel {
 		return diseaseHistoryMapper.getMaxId_SalesModel();
 	}
 
-	public void add(DiseaseHistoryVO diseaseHistoryVO) {
-		diseaseHistoryMapper.insert_SalesModel(diseaseHistoryVO);
+	public void add(DiseaseHistory diseaseHistory) {
+		diseaseHistoryMapper.insert_SalesModel(diseaseHistory.getVO());
 	}
 
-	public void update(DiseaseHistoryVO diseaseHistoryVO) {
-		diseaseHistoryMapper.update(diseaseHistoryVO);
+	public void update(DiseaseHistory diseaseHistory) {
+		diseaseHistoryMapper.update(diseaseHistory.getVO());
 	}
 
 	public void delete(int id) {
+		if (getById(id) == null) return;
 		diseaseHistoryMapper.deleteById(id);
 	}
 }
